updated UI Framework
