+++ src/SFA.DAS.UI.Framework/DriverLocationConfig.cs
@@ -1,7 +1,6 @@
-﻿namespace SFA.DAS.UI.Framework
+﻿namespace SFA.DAS.UI.Framework;
+
+public class DriverLocationConfig
 {
-    public class DriverLocationConfig
-    {
-        public (string chromeWebDriver, string geckoWebDriver, string iEWebDriver, string edgeWebDriver) DriverLocation { get; internal set; }
-    }
+    public (string chromeWebDriver, string geckoWebDriver, string edgeWebDriver) DriverLocation { get; internal set; }
 }
