+++ src/SFA.DAS.UI.Framework/Hooks/AfterScenario/BrowserStackTearDown.cs
@@ -1,16 +1,15 @@
 ﻿using SFA.DAS.UI.Framework.TestSupport;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.UI.Framework.Hooks.AfterScenario
+namespace SFA.DAS.UI.Framework.Hooks.AfterScenario;
+
+[Binding]
+public class BrowserStackTearDown
 {
-    [Binding]
-    public class BrowserStackTearDown
-    {
-        private readonly ScenarioContext _context;
+    private readonly ScenarioContext _context;
 
-        public BrowserStackTearDown(ScenarioContext context) => _context = context;
+    public BrowserStackTearDown(ScenarioContext context) => _context = context;
 
-        [AfterScenario(Order = 12)]
-        public void InformBrowserStackOnFailure() => new BrowserStackTearDownHelper(_context).MarkTestStatus();
-    }
+    [AfterScenario(Order = 12)]
+    public void InformBrowserStackOnFailure() => new BrowserStackTearDownHelper(_context).MarkTestStatus();
 }
