+++ src/SFA.DAS.UI.Framework/Hooks/AfterScenario/DisposeWebDriverTeardown.cs
@@ -1,16 +1,15 @@
 ﻿using SFA.DAS.UI.Framework.TestSupport;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.UI.Framework.Hooks.AfterScenario
+namespace SFA.DAS.UI.Framework.Hooks.AfterScenario;
+
+[Binding]
+public class DisposeWebDriverTeardown
 {
-    [Binding]
-    public class DisposeWebDriverTeardown
-    {
-        private readonly ScenarioContext _context;
+    private readonly ScenarioContext _context;
 
-        public DisposeWebDriverTeardown(ScenarioContext context) => _context = context;
+    public DisposeWebDriverTeardown(ScenarioContext context) => _context = context;
 
-        [AfterScenario(Order = 19)]
-        public void DisposeWebDriver() => new DisposeWebDriverTeardownHelper(_context).DisposeWebDriver();
-    }
+    [AfterScenario(Order = 19)]
+    public void DisposeWebDriver() => new DisposeWebDriverTeardownHelper(_context).DisposeWebDriver();
 }
