+++ src/SFA.DAS.UI.Framework/Hooks/BeforeScenario/UIFrameworkHelpersSetup.cs
@@ -1,28 +1,27 @@
 ﻿using SFA.DAS.UI.FrameworkHelpers;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.UI.Framework.Hooks.BeforeScenario
+namespace SFA.DAS.UI.Framework.Hooks.BeforeScenario;
+
+[Binding]
+public class UIFrameworkHelpersSetup
 {
-    [Binding]
-    public class UIFrameworkHelpersSetup
+    private readonly ScenarioContext _context;
+    private readonly TestSupport.UIFrameworkHelpersSetup _helpersSetup;
+    private readonly FrameworkConfig _config;
+
+    public UIFrameworkHelpersSetup(ScenarioContext context)
     {
-        private readonly ScenarioContext _context;
-        private readonly TestSupport.UIFrameworkHelpersSetup _helpersSetup;
-        private readonly FrameworkConfig _config;
+        _context = context;
+        _config = context.Get<FrameworkConfig>();
+        _helpersSetup = new TestSupport.UIFrameworkHelpersSetup(context);
+    }
 
-        public UIFrameworkHelpersSetup(ScenarioContext context)
-        {
-            _context = context;
-            _config = context.Get<FrameworkConfig>();
-            _helpersSetup = new TestSupport.UIFrameworkHelpersSetup(context);
-        }
+    [BeforeScenario(Order = 4)]
+    public void SetUpUIFrameworkHelpers()
+    {
+        _helpersSetup.SetupUIFrameworkHelpers(false);
 
-        [BeforeScenario(Order = 4)]
-        public void SetUpUIFrameworkHelpers()
-        {
-            _helpersSetup.SetupUIFrameworkHelpers(false);
-
-            _context.Set(new BrowserStackReportingService(_config.BrowserStackSetting));
-        }
+        _context.Set(new BrowserStackReportingService(_config.BrowserStackSetting));
     }
 }
