+++ src/SFA.DAS.UI.Framework/TestSupport/RestartWebDriverHelper.cs
@@ -2,47 +2,46 @@
 using OpenQA.Selenium;
 using SFA.DAS.FrameworkHelpers;
 
-namespace SFA.DAS.UI.Framework.TestSupport
+namespace SFA.DAS.UI.Framework.TestSupport;
+
+public class RestartWebDriverHelper
 {
-    public class RestartWebDriverHelper
+    private readonly BrowserStackTearDownHelper _browserStackTearDownHelper;
+    private readonly DisposeWebDriverTeardownHelper _disposeWebDriverTeardownHelper;
+    private readonly WebDriverSetupHelper _webDriverSetupHelper;
+    private readonly UIFrameworkHelpersSetup _frameworkHelpersSetup;
+    private readonly ObjectContext _objectContext;
+
+    public RestartWebDriverHelper(ScenarioContext context)
     {
-        private readonly BrowserStackTearDownHelper _browserStackTearDownHelper;
-        private readonly DisposeWebDriverTeardownHelper _disposeWebDriverTeardownHelper;
-        private readonly WebDriverSetupHelper _webDriverSetupHelper;
-        private readonly UIFrameworkHelpersSetup _frameworkHelpersSetup;
-        private readonly ObjectContext _objectContext;
+        _objectContext = context.Get<ObjectContext>();
+        _browserStackTearDownHelper = new BrowserStackTearDownHelper(context);
+        _disposeWebDriverTeardownHelper = new DisposeWebDriverTeardownHelper(context);
+        _webDriverSetupHelper = new WebDriverSetupHelper(context);
+        _frameworkHelpersSetup = new UIFrameworkHelpersSetup(context);
+    }
 
-        public RestartWebDriverHelper(ScenarioContext context)
-        {
-            _objectContext = context.Get<ObjectContext>();
-            _browserStackTearDownHelper = new BrowserStackTearDownHelper(context);
-            _disposeWebDriverTeardownHelper = new DisposeWebDriverTeardownHelper(context);
-            _webDriverSetupHelper = new WebDriverSetupHelper(context);
-            _frameworkHelpersSetup = new UIFrameworkHelpersSetup(context);
-        }
+    public void RestartWebDriver(string url, string applicationName)
+    {
+        _browserStackTearDownHelper.MarkTestStatus();
 
-        public void RestartWebDriver(string url, string applicationName)
-        {
-            _browserStackTearDownHelper.MarkTestStatus();
+        _browserStackTearDownHelper.UpdateTestName(_objectContext.GetCurrentApplicationName());
 
-            _browserStackTearDownHelper.UpdateTestName(_objectContext.GetCurrentApplicationName());
+        var webDriver = RestartWebDriver();
 
-            var webDriver = RestartWebDriver();
+        webDriver.Navigate().GoToUrl(url);
 
-            webDriver.Navigate().GoToUrl(url);
+        _objectContext.SetCurrentApplicationName(applicationName);
+    }
 
-            _objectContext.SetCurrentApplicationName(applicationName);
-        }
+    public IWebDriver RestartWebDriver()
+    {
+        _disposeWebDriverTeardownHelper.DisposeWebDriver();
 
-        public IWebDriver RestartWebDriver()
-        {
-            _disposeWebDriverTeardownHelper.DisposeWebDriver();
+        var webdriver = _webDriverSetupHelper.SetupWebDriver();
 
-            var webdriver = _webDriverSetupHelper.SetupWebDriver();
+        _frameworkHelpersSetup.SetupUIFrameworkHelpers(true);
 
-            _frameworkHelpersSetup.SetupUIFrameworkHelpers(true);
-
-            return webdriver;
-        }
+        return webdriver;
     }
 }
