+++ src/SFA.DAS.UI.Framework/TestSupport/SqlHelpers/UsersSqlDataHelper.cs
@@ -1,14 +1,13 @@
 ﻿using SFA.DAS.ConfigurationBuilder;
 using SFA.DAS.FrameworkHelpers;
 
-namespace SFA.DAS.UI.Framework.TestSupport.SqlHelpers
+namespace SFA.DAS.UI.Framework.TestSupport.SqlHelpers;
+
+public class UsersSqlDataHelper : SqlDbHelper
 {
-    public class UsersSqlDataHelper : SqlDbHelper
-    {
-        public UsersSqlDataHelper(DbConfig dbConfig) : base(dbConfig.UsersDbConnectionString) { }
+    public UsersSqlDataHelper(DbConfig dbConfig) : base(dbConfig.UsersDbConnectionString) { }
 
-        public void ReinstateAccountInDb(string email) => ExecuteSqlCommand($"UPDATE [User] SET IsSuspended = 0, LastSuspendedDate = null WHERE Email = '{email}'");
+    public void ReinstateAccountInDb(string email) => ExecuteSqlCommand($"UPDATE [User] SET IsSuspended = 0, LastSuspendedDate = null WHERE Email = '{email}'");
 
-        public string GetUserId(string email) => GetDataAsString($"SELECT Id FROM dbo.[User] WHERE email = '{email}'");
-    }
+    public string GetUserId(string email) => GetDataAsString($"SELECT Id FROM dbo.[User] WHERE email = '{email}'");
 }
