+++ src/SFA.DAS.UI.Framework/TestSupport/WebdriverAddCapabilities.cs
@@ -3,36 +3,35 @@ using SFA.DAS.FrameworkHelpers;
 using System.Collections.Generic;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.UI.Framework.TestSupport
+namespace SFA.DAS.UI.Framework.TestSupport;
+
+public abstract class WebdriverAddCapabilities
 {
-    public abstract class WebdriverAddCapabilities
+    protected readonly ObjectContext objectContext;
+
+    public WebdriverAddCapabilities(ScenarioContext context) => objectContext = context.Get<ObjectContext>();
+
+    protected void AddChromeCapabilities(IWebDriver webDriver)
     {
-        protected readonly ObjectContext objectContext;
+        var cap = AddBrowserCapabilities(webDriver);
 
-        public WebdriverAddCapabilities(ScenarioContext context) => objectContext = context.Get<ObjectContext>();
+        foreach (var item in cap["chrome"] as Dictionary<string, object>) objectContext.Replace(item.Key, item.Value);
+    }
 
-        protected void AddChromeCapabilities(IWebDriver webDriver)
-        {
-            var cap = AddBrowserCapabilities(webDriver);
+    protected void AddEdgeCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
 
-            foreach (var item in cap["chrome"] as Dictionary<string, object>) objectContext.Replace(item.Key, item.Value);
-        }
+    protected void AddFireFoxCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
 
-        protected void AddEdgeCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
+    protected void AddIeCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
 
-        protected void AddFireFoxCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
+    private ICapabilities AddBrowserCapabilities(IWebDriver webDriver)
+    {
+        var cap = (webDriver as WebDriver).Capabilities;
 
-        protected void AddIeCapabilities(IWebDriver webDriver) => AddBrowserCapabilities(webDriver);
+        objectContext.SetBrowserName(cap["browserName"]);
 
-        private ICapabilities AddBrowserCapabilities(IWebDriver webDriver)
-        {
-            var cap = (webDriver as WebDriver).Capabilities;
+        objectContext.SetBrowserVersion(cap["browserVersion"]);
 
-            objectContext.SetBrowserName(cap["browserName"]);
-
-            objectContext.SetBrowserVersion(cap["browserVersion"]);
-
-            return cap;
-        }
+        return cap;
     }
 }
